# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Document {
  id: number;
  fileName: string;
  fileSize: number;
  status: DocumentStatus;
  /** @nullable */
  pageCount: number | null;
  uploadedAt: string;
  /** @nullable */
  analyzedAt: string | null;
}

export interface DocumentStats {
  total: number;
  pending: number;
  processing: number;
  completed: number;
  failed: number;
  /** @nullable */
  avgConfidence: number | null;
}

export interface ExtractedParty {
  role: string;
  name: string;
  confidence: number;
}

export interface ExtractedDirective {
  type: string;
  description: string;
  sourceText: string;
  confidence: number;
}

export interface HighlightedSection {
  label: string;
  text: string;
  confidence: number;
  category: string;
}

export interface DocumentAnalysis {
  /** @nullable */
  caseNumber: string | null;
  /** @nullable */
  caseNumberConfidence: number | null;
  /** @nullable */
  caseName: string | null;
  /** @nullable */
  caseNameConfidence: number | null;
  /** @nullable */
  courtName: string | null;
  /** @nullable */
  courtNameConfidence: number | null;
  /** @nullable */
  judgmentDate: string | null;
  /** @nullable */
  judgmentDateConfidence: number | null;
  /** @nullable */
  judge: string | null;
  /** @nullable */
  judgeConfidence: number | null;
  parties: ExtractedParty[];
  directives: ExtractedDirective[];
  highlights: HighlightedSection[];
  /** @nullable */
  summary: string | null;
  /** @nullable */
  overallConfidence: number | null;
}

export type DocumentWithAnalysisStatus =
  (typeof DocumentWithAnalysisStatus)[keyof typeof DocumentWithAnalysisStatus];

export const DocumentWithAnalysisStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface DocumentWithAnalysis {
  id: number;
  fileName: string;
  fileSize: number;
  status: DocumentWithAnalysisStatus;
  /** @nullable */
  pageCount: number | null;
  uploadedAt: string;
  /** @nullable */
  analyzedAt: string | null;
  analysis: DocumentAnalysis | null;
}

export type ListDocumentsParams = {
  status?: ListDocumentsStatus;
};

export type ListDocumentsStatus =
  (typeof ListDocumentsStatus)[keyof typeof ListDocumentsStatus];

export const ListDocumentsStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;