# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: number;
  fileName: string;
  fileSize: number;
  status: DocumentStatus;
  /** @nullable */
  pageCount: number | null;
  uploadedAt: Date;
  /** @nullable */
  analyzedAt: Date | null;
}