# lib/api-zod/src/generated/types/documentAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedDirective } from "./extractedDirective";
import type { ExtractedParty } from "./extractedParty";
import type { HighlightedSection } from "./highlightedSection";

export interface DocumentAnalysis {
  /** @nullable */
  caseNumber: string | null;
  /** @nullable */
  caseNumberConfidence: number | null;
  /** @nullable */
  caseName: string | null;
  /** @nullable */
  caseNameConfidence: number | null;
  /** @nullable */
  courtName: string | null;
  /** @nullable */
  courtNameConfidence: number | null;
  /** @nullable */
  judgmentDate: string | null;
  /** @nullable */
  judgmentDateConfidence: number | null;
  /** @nullable */
  judge: string | null;
  /** @nullable */
  judgeConfidence: number | null;
  parties: ExtractedParty[];
  directives: ExtractedDirective[];
  highlights: HighlightedSection[];
  /** @nullable */
  summary: string | null;
  /** @nullable */
  overallConfidence: number | null;
}