# lib/api-zod/src/generated/types/documentStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentStats {
  total: number;
  pending: number;
  processing: number;
  completed: number;
  failed: number;
  /** @nullable */
  avgConfidence: number | null;
}