# lib/api-zod/src/generated/types/documentWithAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentAnalysis } from "./documentAnalysis";
import type { DocumentWithAnalysisStatus } from "./documentWithAnalysisStatus";

export interface DocumentWithAnalysis {
  id: number;
  fileName: string;
  fileSize: number;
  status: DocumentWithAnalysisStatus;
  /** @nullable */
  pageCount: number | null;
  uploadedAt: Date;
  /** @nullable */
  analyzedAt: Date | null;
  analysis: DocumentAnalysis | null;
}