# lib/api-zod/src/generated/types/extractedParty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Understanding & Extraction Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedParty {
  role: string;
  name: string;
  confidence: number;
}